major project/requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
streamlit
joblib
plotly
fpdf2
bcrypt
openai
python-dotenv
